Post Progress check + added file upload fix/ISPJ/AdminApp/app/requirements.txt
annotated-types==0.7.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.18
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
limiter==0.5.0
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
plotly==5.24.1
prometheus_client==0.21.1
prometheus_flask_exporter==0.23.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyotp==2.9.0
python-dotenv==1.0.1
python-json-logger==3.2.1
requests==2.32.3
rich==13.9.4
SQLAlchemy==2.0.36
StrEnum==0.4.15
tenacity==9.0.0
token-bucket==0.3.0
typing_extensions==4.12.2
urllib3==2.2.3
vonage==4.1.2
vonage-account==1.1.1
vonage-application==2.0.1
vonage-http-client==1.5.0
vonage-jwt==1.1.5
vonage-messages==1.3.0
vonage-network-auth==1.0.2
vonage-network-number-verification==1.0.2
vonage-network-sim-swap==1.1.2
vonage-number-insight==1.0.6
vonage-numbers==1.0.4
vonage-sms==1.1.5
vonage-subaccounts==1.0.4
vonage-users==1.2.1
vonage-utils==1.1.4
vonage-verify==2.1.0
vonage-verify-legacy==1.0.1
vonage-video==1.0.4
vonage-voice==1.1.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1

#Kenzie requirements:
streamlit
pandasai
numpy==1.23.5
pandas==1.5.3
email-validator
pycryptodome